reserialization

+++ tests/test.py
@@ -1,19 +1,61 @@
+#!/usr/bin/env python3
+# I'm not including any of the unedited files with this package so we are just not going to export the tests
+
+import os
+import unittest
+
 import ainb
 import ainb.graph
 
-ainb.AINB.from_file("tests/data/ChuchuFire.action.root.ainb").save_json("tests/output")
-ainb.AINB.from_file("tests/data/Alerted.actionseq.root.ainb").save_json("tests/output")
-ainb.AINB.from_file("tests/data/Main.module.ainb").save_json("tests/output")
-ainb.AINB.from_file("tests/data/Bird.action.fleetothesky.module.ainb").save_json("tests/output")
-ainb.AINB.from_file("tests/data/DungeonBossRito.action.SpreadShootAttack.module.ainb").save_json("tests/output")
-ainb.AINB.from_file("tests/data/Drake.sp.action.root.ainb").save_json("tests/output")
+def fix_path(path: str) -> str:
+    return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
 
-orig: ainb.AINB = ainb.AINB.from_file("tests/data/Drake.sp.action.root.ainb")
-new: ainb.AINB = ainb.AINB.from_json_text(orig.to_json())
+INPUT_DIRECTORY: str = fix_path("data")
 
-assert orig.as_dict() == new.as_dict(), "oops they don't match"
+class RoundtripTest(unittest.TestCase):
+    def test_json_roundtrip(self) -> None:
+        ainb.set_splatoon3()
+        for file in os.listdir(INPUT_DIRECTORY):
+            if not os.path.isfile(os.path.join(INPUT_DIRECTORY, file)):
+                continue
+            try:
+                print(file)
+                orig: ainb.AINB = ainb.AINB.from_file(os.path.join(INPUT_DIRECTORY, file), read_only=False)
+                new: ainb.AINB = ainb.AINB.from_json_text(orig.to_json())
+            except Exception as e:
+                self.fail(f"{file} failed: {e.args}")
+            self.assertDictEqual(orig.as_dict(), new.as_dict(), f"{file} is mismatching")
+    
+    def test_ainb_roundtrip(self) -> None:
+        ainb.set_splatoon3()
+        for file in os.listdir(INPUT_DIRECTORY):
+            if not os.path.isfile(os.path.join(INPUT_DIRECTORY, file)):
+                continue
+            try:
+                print(file)
+                orig: ainb.AINB = ainb.AINB.from_file(os.path.join(INPUT_DIRECTORY, file), read_only=False)
+                new: ainb.AINB = ainb.AINB.from_binary(orig.to_binary())
+            except Exception as e:
+                self.fail(f"{file} failed: {e.args}")
+            self.assertDictEqual(orig.as_dict(), new.as_dict(), f"{file} is mismatching")
 
-ainb.graph.graph_all_commands(ainb.AINB.from_file("tests/data/Drake.sp.action.root.ainb"), render=True, output_dir="tests/output")
-ainb.graph.graph_all_commands(ainb.AINB.from_file("tests/data/DungeonBossRito.action.SpreadShootAttack.module.ainb"), render=True, output_dir="tests/output")
-ainb.graph.graph_all_commands(ainb.AINB.from_file("tests/data/Main.module.ainb"), render=True, output_dir="tests/output")
-ainb.graph.graph_all_commands(ainb.AINB.from_json("tests/output/SplPlayerMake.root.json"), render=True, output_dir="tests/output")
+class GraphTest(unittest.TestCase):
+    def test(self) -> None:
+        ainb.set_splatoon3()
+        for file in os.listdir(INPUT_DIRECTORY):
+            if not os.path.isfile(os.path.join(INPUT_DIRECTORY, file)):
+                continue
+            try:
+                if ".logic" in file:
+                    ainb.graph.graph_all_nodes(ainb.AINB.from_file(os.path.join(INPUT_DIRECTORY, file), read_only=False))
+                else:
+                    ainb.graph.graph_all_nodes(ainb.AINB.from_file(os.path.join(INPUT_DIRECTORY, file), read_only=False))
+            except Exception as e:
+                self.fail(f"Failed to graph {file}: {e.args}")
+
+if __name__ == "__main__":
+    import sys
+    if len(sys.argv) > 1:
+        INPUT_DIRECTORY = sys.argv[1]
+        sys.argv = sys.argv[0:1] + sys.argv[2:]
+    unittest.main()
